Evento y notificacion correo modificados
