+++ src/layouts/dashboard/header.jsx
@@ -16,10 +16,9 @@ import Logo from 'src/components/logoMini';
 import SvgColor from 'src/components/svg-color';
 import { useSettingsContext } from 'src/components/settings';
 
-import PendingModalUser from 'src/sections/calendariobeneficiario/pendingModalUser';
+// import PendingModalUser from 'src/sections/calendariobeneficiario/pendingModalUser';
 import PendingModal from 'src/sections/overview/calendarioespecialista/view/pendingModal';
 
-import Searchbar from '../common/searchbar';
 import { NAV, HEADER } from '../config-layout';
 import SettingsButton from '../common/settings-button';
 import AccountPopover from '../common/account-popover';
